requirements.txt
# === Web Framework ===
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
# === Database ===
sqlalchemy==2.0.35
alembic==1.13.3
# === LLM & Agent ===
openai==1.54.0
langchain==0.3.7
langchain-openai==0.2.6
langgraph==0.2.45
langchain-community==0.3.5
tiktoken==0.8.0
# === RAG / Vector DB ===
chromadb==0.5.18
langchain-chroma==0.1.4
# === Ingestion (attachments) ===
pypdf==5.1.0
pillow==11.0.0
pytesseract==0.3.13
beautifulsoup4==4.12.3
requests==2.32.3
# === Scheduling ===
apscheduler==3.10.4
# === Configuration & Utilities ===
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
loguru==0.7.2
# === Frontend (later phase) ===
streamlit==1.40.0
# === Dev/Testing ===
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pandas==2.2.3              # for cost dashboard tables